requirements.txt
Pillow
pyfiglet
keyboard